NLP/UNIMO-2/requirements.txt
astor==0.8.1
astroid==2.11.2
black==22.3.0
cached-property==1.5.2
certifi==2020.12.5
chardet==4.0.0
click==8.1.1
colorama==0.4.4
colorlog==5.0.1
cycler==0.10.0
decorator==4.4.2
dill==0.3.4
gast==0.4.0
h5py==3.3.0
idna==2.10
imageio==2.9.0
importlib-metadata==4.11.3
isort==5.10.1
jieba==0.42.1
joblib==1.0.1
kiwisolver==1.3.1
lazy-object-proxy==1.7.1
matplotlib==3.4.2
mccabe==0.7.0
multiprocess==0.70.12.2
mypy-extensions==0.4.3
networkx==2.5.1
numpy==1.20.2
opencv-python==4.5.2.52
paddlecloudplatform==3.0.13
paddlenlp==2.0.7
paddlepaddle-gpu==2.1.2
pathspec==0.9.0
Pillow==8.2.0
platformdirs==2.5.1
protobuf==3.15.8
pylint==2.13.2
pyparsing==2.4.7
pyrouge==0.1.3
python-dateutil==2.8.1
PyWavelets==1.1.1
regex==2021.4.4
requests==2.25.1
scikit-image==0.18.1
scikit-learn==0.24.2
scipy==1.7.0
sentencepiece==0.1.96
seqeval==1.2.2
six==1.15.0
termcolor==1.1.0
threadpoolctl==2.2.0
tifffile==2021.6.14
tomli==2.0.1
typed-ast==1.5.2
typing-extensions==4.1.1
urllib3==1.26.4
wrapt==1.14.0
xlrd==1.2.0
xlwt==1.2.0
zipp==3.7.0
